Extract file setup logic to StudentTest class to reduce code duplication.
Extract group related methods into GroupBrowserPage.
