+++ apps/student/src/androidTest/java/com/instructure/student/ui/interaction/ShareExtensionInteractionTest.kt
@@ -16,13 +16,9 @@
  */
 package com.instructure.student.ui.interaction
 
-import android.app.Activity
-import android.app.Instrumentation
 import android.content.Intent
 import android.net.Uri
-import androidx.core.content.FileProvider
 import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
 import androidx.test.uiautomator.UiDevice
@@ -33,11 +29,9 @@ import com.instructure.canvas.espresso.mockCanvas.addAssignment
 import com.instructure.canvas.espresso.mockCanvas.init
 import com.instructure.canvasapi2.models.Assignment
 import com.instructure.canvasapi2.models.User
-import com.instructure.pandautils.utils.Const
 import com.instructure.student.ui.utils.StudentTest
 import com.instructure.student.ui.utils.tokenLogin
 import dagger.hilt.android.testing.HiltAndroidTest
-import org.hamcrest.core.AllOf
 import org.junit.Test
 import java.io.File
 
@@ -277,20 +271,6 @@ class ShareExtensionInteractionTest : StudentTest() {
         tokenLogin(MockCanvas.data.domain, token!!, student)
     }
 
-    private fun setupFileOnDevice(fileName: String): Uri {
-        copyAssetFileToExternalCache(activityRule.activity, fileName)
-
-        val dir = activityRule.activity.externalCacheDir
-        val file = File(dir?.path, fileName)
-
-        val instrumentationContext = InstrumentationRegistry.getInstrumentation().context
-        return FileProvider.getUriForFile(
-            instrumentationContext,
-            "com.instructure.candroid" + Const.FILE_PROVIDER_AUTHORITY,
-            file
-        )
-    }
-
     private fun shareExternalFile(uri: Uri) {
         val intent = Intent().apply {
             action = Intent.ACTION_SEND
@@ -315,24 +295,5 @@ class ShareExtensionInteractionTest : StudentTest() {
         chooser.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
 
         InstrumentationRegistry.getInstrumentation().context.startActivity(chooser)
-    }
-
-    private fun stubFilePickerIntent(fileName: String) {
-        val resultData = Intent()
-        val dir = activityRule.activity.externalCacheDir
-        val file = File(dir?.path, fileName)
-        val newFileUri = FileProvider.getUriForFile(
-            activityRule.activity,
-            "com.instructure.candroid" + Const.FILE_PROVIDER_AUTHORITY,
-            file
-        )
-        resultData.data = newFileUri
-
-        Intents.intending(
-            AllOf.allOf(
-                IntentMatchers.hasAction(Intent.ACTION_GET_CONTENT),
-                IntentMatchers.hasType("*/*"),
-            )
-        ).respondWith(Instrumentation.ActivityResult(Activity.RESULT_OK, resultData))
     }
 }
